85. Testing Spring MVC Web Controllers - Error Page for Delete Student

+++ 4-springbootmvc-unit-project/src/main/java/com/luv2code/springmvc/controller/GradebookController.java
@@ -48,6 +48,12 @@ public class GradebookController {
     @GetMapping("/delete/student/{id}")
     public String deleteStudent(@PathVariable int id, Model m)
     {
+
+        if(!studentService.checkIfStudentIsNull(id))
+        {
+            return "error";
+        }
+
         studentService.deleteStudent(id);
         Iterable<CollegeStudent> collegeStudents = studentService.getGradebook();
         m.addAttribute("students",collegeStudents);
